+++ owWMO/WMO_Light.h
@@ -21,7 +21,7 @@ struct WMOLightDef
 	float unk[4];
 
 	//
-	const static size_t __size = 48;
+	const static uint32_t __size = 48;
 };
 
 class WMOLight
